backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.29.0

# ── Database ──────────────────────────────────────────────────────────────────
# sqlalchemy[asyncio] installs the greenlet dependency required for async use.
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1

# ── Data validation & settings ────────────────────────────────────────────────
pydantic[email]==2.6.4
pydantic-settings==2.2.1

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── AI ────────────────────────────────────────────────────────────────────────
anthropic==0.25.0

# ── Real-time (WebSocket pub/sub) ─────────────────────────────────────────────
# redis[asyncio] bundles hiredis for fast response parsing.
redis[asyncio]==5.0.3

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
